fixed reports delete error

+++ lib/reports.ts
@@ -1,3 +1,4 @@
+"use server";
 import { connectToDB } from "@utils/database";
 import TripCommentReport, {
   ITripCommentReport,
@@ -5,8 +6,7 @@ import TripCommentReport, {
 } from "@models/tripCommentReport";
 import { revalidatePath } from "next/cache";
 import TripComment from "@models/tripComment";
-
-export const REPORTS_PER_PAGE = 5;
+import { REPORTS_PER_PAGE } from "@app/dashboard/reports/page";
 export const getReports = async (q: string | RegExp, page: number) => {
   const regex = new RegExp(q, "i");
   try {
@@ -22,11 +22,7 @@ export const getReports = async (q: string | RegExp, page: number) => {
       .limit(REPORTS_PER_PAGE)
       .skip(REPORTS_PER_PAGE * (page - 1))
       .lean();
-    console.log(reports);
-    reports.forEach((report) => {
-      report._id = report._id.toString("base64");
-      report.tripCommentId._id = report.tripCommentId._id.toString("base64");
-    });
+
     return { count, reports };
   } catch (err) {
     throw new Error("Failed to fetch reports!");
@@ -38,16 +34,25 @@ export const deleteTripCommentReport = async (
 ) => {
   const { tripCommentId, id } = Object.fromEntries(formData);
 
+  console.log("commentIdObject", tripCommentId);
   try {
-    connectToDB();
-    await TripComment.findByIdAndDelete(tripCommentId);
-    await TripCommentReport.findByIdAndDelete(id);
+    await connectToDB(); // Ensure the database connection is awaited
+    const deletedComment = await TripComment.findByIdAndDelete(tripCommentId);
+    if (!deletedComment) {
+      throw new Error(`Failed to delete TripComment with ID: ${tripCommentId}`);
+    }
+    console.log("hey");
+
+    const deletedReport = await TripCommentReport.findByIdAndDelete(id);
+    if (!deletedReport) {
+      throw new Error(`Failed to delete TripCommentReport with ID: ${id}`);
+    }
+
+    revalidatePath("/dashboard/reports");
   } catch (err) {
     console.log(err);
     throw new Error("Failed to delete comment!");
   }
-
-  revalidatePath("/dashboard/reports");
 };
 
 export const addReport = async (
